add SG symbol for loading image

+++ src/components/LoginComponents/signUp/SignUp2.js
@@ -6,12 +6,13 @@ import PropTypes from 'prop-types'
 import { styles, SignText, Tcpp } from './SignUp'
 
 import deviceStorage from '../../../services/deviceStorage'
-import { JWT, FNAME, LNAME, DOB } from '../../../constants'
+import { validateEmail } from '../../../helpers/validators'
+import { JWT, FNAME, LNAME, DOB, EMAIL_REQUIRED, EMAIL_MALFORMED, PASSWORD_REQUIRED, PASSWORD_ERROR_CONFIRM } from '../../../constants'
 
 // GQL
 const signupMutation = gql`
-  mutation SignUp($email: String!, $password: String!, $firstName: String!, $lastName: String!) {
-    signup(input: {email: $email, password: $password, firstName: $firstName, lastName: $lastName}) {
+  mutation SignUp($email: String!, $password: String!, $firstName: String!, $lastName: String!, $dateOfBirth: Date!) {
+    signup(input: {email: $email, password: $password, firstName: $firstName, lastName: $lastName, dateOfBirth: $dateOfBirth}) {
       token
     }
   }
@@ -21,8 +22,11 @@ const signupMutation = gql`
 export class SignUp2 extends Component {
   state = {
     email: '',
+    emailError: '',
     password: '',
+    passwordError: '',
     confirmPass: '',
+    confirmPassError: '',
     hidePassword: true
   }
 
@@ -38,47 +42,65 @@ export class SignUp2 extends Component {
   }
 
   handleEmailChange = (email) => {
-    this.setState({ email })
+    this.setState({ email, emailError: '' })
   }
 
   handlePasswordChange = (password) => {
-    this.setState({ password })
+    this.setState({ password, passwordError: '' })
   }
 
   handleConfirmPassChange = (confirmPass) => {
-    this.setState({ confirmPass })
+    this.setState({ confirmPass, confirmPassError: '' })
   }
 
   handleSignUp = async () => {
-    const { email, password, confirmPass, } = this.state
+    let { email, password, confirmPass } = this.state
+
+    email = email.toLowerCase()
+
+    //handle validation
+    if (!email) {
+      return this.setState({emailError: EMAIL_REQUIRED})
+    } else if (!validateEmail(email)) {
+      return this.setState({emailError: EMAIL_MALFORMED})
+    } else if (!password) {
+      return this.setState({passwordError: PASSWORD_REQUIRED})
+    } else if (!confirmPass || password !== confirmPass) {
+      return this.setState({confirmPassError: PASSWORD_ERROR_CONFIRM})
+    }
+
     let firstName = await deviceStorage.getItem(FNAME)
     let lastName = await deviceStorage.getItem(LNAME)
-    let dob = await deviceStorage.getItem(DOB)
-    console.log(this.props)
-    this.props.signup(email, password, firstName, lastName)
+
+    // Stored in AsyncStorage as string
+    let dateOfBirth = await deviceStorage.getItem(DOB)
+    dateOfBirth = Number(dateOfBirth)
+
+    this.props.signup(email, password, firstName, lastName, dateOfBirth)
       .then(async (response) => {
         let { data } = response
-        console.log(data)
         await deviceStorage.saveItem(JWT, data.signup.token)
         await deviceStorage.deleteMultipleItems([FNAME,LNAME,DOB])
+        
+        // update App state with jwt and rerender
         this.props.screenProps.saveJWT(data.signup.token)
       }).catch((error) => {
         console.log(error)
       })
-
-
-    console.log(email, password, confirmPass, firstName, lastName, dob)
   }
 
   render() {
-    const { email, password, confirmPass } = this.state
+    const { email, password, confirmPass, emailError, passwordError, confirmPassError } = this.state
     return (
       <TouchableWithoutFeedback onPress={Keyboard.dismiss}>
         <View style={styles.container}>
           <View>
             <View style={[styles.row, styles.logInButtonWrapper]}>
+            
+              <View style={styles.row}/>
+
               <TextInput
-              style={[styles.row, styles.textInput]}
+              style={[styles.textInput]}
               onChangeText={email => this.handleEmailChange(email)}
               value={this.state.email}
               placeholder="EMAIL"
@@ -88,9 +110,14 @@ export class SignUp2 extends Component {
                 this.refs.password.focus();
               }}
               />
+              <View>
+                <Text>{emailError}</Text>
+              </View>
+              <View style={styles.row}/>
+
               <TextInput
               ref='password'
-              style={[styles.row, styles.textInput]}
+              style={[styles.textInput]}
               onChangeText={password => this.handlePasswordChange(password)}
               value={this.state.password}
               placeholder="PASSWORD"
@@ -101,9 +128,14 @@ export class SignUp2 extends Component {
                 this.refs.confirmPass.focus();
               }}
               />
+              <View>
+                <Text>{passwordError}</Text>
+              </View>
+              <View style={styles.row}/>
+
               <TextInput
               ref='confirmPass'
-              style={[styles.row, styles.textInput]}
+              style={[styles.textInput]}
               onChangeText={confirmPass => this.handleConfirmPassChange(confirmPass)}
               value={this.state.confirmPass}
               placeholder="CONFIRM PASSWORD"
@@ -111,6 +143,9 @@ export class SignUp2 extends Component {
               returnKeyType="done"
               secureTextEntry = { this.state.hidePassword }
               />
+              <View>
+                <Text>{confirmPassError}</Text>
+              </View>
               <View style={styles.row}/>
 
               <View style={[styles.row, styles.logInButtonWrapper]}>
@@ -133,7 +168,7 @@ export default graphql (
   signupMutation,
   {
     props: ({ mutate }) => ({
-      signup: (email, password, firstName, lastName) => mutate({ variables: { email, password, firstName, lastName } })
+      signup: (email, password, firstName, lastName, dateOfBirth) => mutate({ variables: { email, password, firstName, lastName, dateOfBirth } })
     }),
   },
 )(SignUp2)
